wrap CanUserPerform in a short-lived cache
- We've observed bursts of CanUserPerform messages. So cache the result for about ~10 seconds (and single-flight them by team/subteam name)
