+++ go/teams/service_helper.go
@@ -1242,132 +1242,6 @@ func GetKBFSTeamSettings(ctx context.Context, g *libkb.GlobalContext, isPublic b
 	return res, err
 }
 
-func CanUserPerform(ctx context.Context, g *libkb.GlobalContext, teamname string) (ret keybase1.TeamOperation, err error) {
-	team, err := Load(ctx, g, keybase1.LoadTeamArg{
-		Name:    teamname,
-		StaleOK: true,
-		Public:  false, // assume private team
-	})
-	if err != nil {
-		// Note: we eat the error here, assuming it meant this user
-		// is not a member
-		g.Log.CWarningf(ctx, "CanUserPerform team Load failure, continuing: %v)", err)
-		return ret, nil
-	}
-	meUV, err := team.currentUserUV(ctx)
-	if err != nil {
-		return ret, err
-	}
-
-	isImplicitAdmin := func() (bool, error) {
-		if team.ID.IsRootTeam() {
-			return false, nil
-		}
-		uvs, err := g.GetTeamLoader().ImplicitAdmins(ctx, team.ID)
-		if err != nil {
-			return false, err
-		}
-		for _, uv := range uvs {
-			if uv == meUV {
-				return true, nil
-			}
-		}
-		return false, nil
-	}
-
-	teamRole, err := team.MemberRole(ctx, meUV)
-	if err != nil {
-		return ret, err
-	}
-
-	isRoleOrAbove := func(role keybase1.TeamRole) bool {
-		return teamRole.IsOrAbove(role)
-	}
-
-	isWriter := func() bool {
-		return isRoleOrAbove(keybase1.TeamRole_WRITER)
-	}
-
-	canMemberShowcase := func() (bool, error) {
-		if err != nil {
-			return false, err
-		}
-		if teamRole.IsOrAbove(keybase1.TeamRole_ADMIN) {
-			return true, nil
-		} else if teamRole == keybase1.TeamRole_NONE {
-			return false, nil
-		}
-		showcase, err := GetTeamShowcase(ctx, g, teamname)
-		if err != nil {
-			return false, err
-		}
-		return showcase.AnyMemberShowcase, nil
-	}
-
-	hasOtherOwner := func() (bool, error) {
-		owners, err := team.UsersWithRole(keybase1.TeamRole_OWNER)
-		if err != nil {
-			return false, err
-		}
-		if len(owners) > 1 {
-			return true, nil
-		}
-		for _, owner := range owners {
-			if owner == meUV {
-				g.Log.CDebugf(ctx, "hasOtherOwner: I am the sole owner")
-				return false, nil
-			}
-		}
-		return true, nil
-	}
-
-	var implicitAdmin bool
-	implicitAdmin, err = isImplicitAdmin()
-	if err != nil {
-		return ret, err
-	}
-	var admin bool
-	admin = isRoleOrAbove(keybase1.TeamRole_ADMIN)
-
-	ret.ManageMembers = admin || implicitAdmin
-	ret.ManageSubteams = admin || implicitAdmin
-	ret.SetTeamShowcase = admin || implicitAdmin
-	ret.ChangeOpenTeam = admin || implicitAdmin
-	ret.ChangeTarsDisabled = admin || implicitAdmin
-
-	ret.ListFirst = implicitAdmin
-	ret.JoinTeam = teamRole == keybase1.TeamRole_NONE && implicitAdmin
-	ret.SetPublicityAny = admin || implicitAdmin
-
-	if teamRole != keybase1.TeamRole_NONE {
-		leaveTeam := true
-		if teamRole == keybase1.TeamRole_OWNER {
-			leaveTeam, err = hasOtherOwner()
-			if err != nil {
-				return ret, err
-			}
-		}
-		ret.LeaveTeam = leaveTeam
-	}
-
-	writer := isWriter()
-	ret.CreateChannel = writer
-
-	ret.SetMemberShowcase, err = canMemberShowcase()
-	if err != nil {
-		return ret, err
-	}
-
-	ret.DeleteChannel = admin
-	ret.RenameChannel = writer
-	ret.EditChannelDescription = writer
-	ret.DeleteChatHistory = admin
-	ret.SetRetentionPolicy = admin
-	ret.Chat = isRoleOrAbove(keybase1.TeamRole_READER)
-
-	return ret, err
-}
-
 func RotateKey(ctx context.Context, g *libkb.GlobalContext, teamID keybase1.TeamID) (err error) {
 	defer g.CTrace(ctx, fmt.Sprintf("RotateKey(%v)", teamID), func() error { return err })()
 	return RetryOnSigOldSeqnoError(ctx, g, func(ctx context.Context, attempt int) error {
